AgentServer/pyproject.toml
[project]
name = "agentserver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.132.0",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.37",
    "livekit-agents[openai,silero,turn-detector]~=1.4",
    "livekit-api>=1.1.0",
    "livekit-plugins-noise-cancellation~=0.2",
    "uvicorn>=0.41.0",
    "psycopg[binary]>=3.3.3",
    "alembic>=1.18.4",
    "pydantic-settings>=2.13.1",
    "psycopg2-binary>=2.9.11",
]
